requirements.txt
Django==1.8.3
-e git://github.com/openhealthcare/opal.git@v0.4.2#egg=opal
mock
python-dateutil==2.4.2
